refactor: VulnerabilityCorrelator class for multi-stage attack chain discovery and update imports
