+++ server_core/vulnerability_correlator.py
@@ -1,0 +1,123 @@
+import logging
+
+class VulnerabilityCorrelator:
+    """Correlate vulnerabilities for multi-stage attack chain discovery"""
+
+    def __init__(self, logger=None):
+        self.logger = logger or logging.getLogger(__name__)
+        self.attack_patterns = {
+            "privilege_escalation": ["local", "kernel", "suid", "sudo"],
+            "remote_execution": ["remote", "network", "rce", "code execution"],
+            "persistence": ["service", "registry", "scheduled", "startup"],
+            "lateral_movement": ["smb", "wmi", "ssh", "rdp"],
+            "data_exfiltration": ["file", "database", "memory", "network"]
+        }
+
+        self.software_relationships = {
+            "windows": ["iis", "office", "exchange", "sharepoint"],
+            "linux": ["apache", "nginx", "mysql", "postgresql"],
+            "web": ["php", "nodejs", "python", "java"],
+            "database": ["mysql", "postgresql", "oracle", "mssql"]
+        }
+
+    def find_attack_chains(self, target_software, max_depth=3):
+        """Find multi-vulnerability attack chains"""
+        try:
+            # This is a simplified implementation
+            # Real version would use graph algorithms and ML
+
+            chains = []
+
+            # Example attack chain discovery logic
+            base_software = target_software.lower()
+
+            # Find initial access vulnerabilities
+            initial_vulns = self._find_vulnerabilities_by_pattern(base_software, "remote_execution")
+
+            for initial_vuln in initial_vulns[:3]:  # Limit for demo
+                chain = {
+                    "chain_id": f"chain_{len(chains) + 1}",
+                    "target": target_software,
+                    "stages": [
+                        {
+                            "stage": 1,
+                            "objective": "Initial Access",
+                            "vulnerability": initial_vuln,
+                            "success_probability": 0.75
+                        }
+                    ],
+                    "overall_probability": 0.75,
+                    "complexity": "MEDIUM"
+                }
+
+                # Find privilege escalation
+                priv_esc_vulns = self._find_vulnerabilities_by_pattern(base_software, "privilege_escalation")
+                if priv_esc_vulns:
+                    chain["stages"].append({
+                        "stage": 2,
+                        "objective": "Privilege Escalation",
+                        "vulnerability": priv_esc_vulns[0],
+                        "success_probability": 0.60
+                    })
+                    chain["overall_probability"] *= 0.60
+
+                # Find persistence
+                persistence_vulns = self._find_vulnerabilities_by_pattern(base_software, "persistence")
+                if persistence_vulns and len(chain["stages"]) < max_depth:
+                    chain["stages"].append({
+                        "stage": 3,
+                        "objective": "Persistence",
+                        "vulnerability": persistence_vulns[0],
+                        "success_probability": 0.80
+                    })
+                    chain["overall_probability"] *= 0.80
+
+                chains.append(chain)
+
+            return {
+                "success": True,
+                "target_software": target_software,
+                "total_chains": len(chains),
+                "attack_chains": chains,
+                "recommendation": self._generate_chain_recommendations(chains)
+            }
+
+        except Exception as e:
+            self.logger.error(f"Error finding attack chains: {str(e)}")
+            return {"success": False, "error": str(e)}
+
+    def _find_vulnerabilities_by_pattern(self, software, pattern_type):
+        """Find vulnerabilities matching attack pattern"""
+        # Simplified mock data - real implementation would query CVE database
+        mock_vulnerabilities = [
+            {
+                "cve_id": "CVE-2024-1234",
+                "description": f"Remote code execution in {software}",
+                "cvss_score": 9.8,
+                "exploitability": "HIGH"
+            },
+            {
+                "cve_id": "CVE-2024-5678",
+                "description": f"Privilege escalation in {software}",
+                "cvss_score": 7.8,
+                "exploitability": "MEDIUM"
+            }
+        ]
+
+        return mock_vulnerabilities
+
+    def _generate_chain_recommendations(self, chains):
+        """Generate recommendations for attack chains"""
+        if not chains:
+            return "No viable attack chains found for target"
+
+        recommendations = [
+            f"Found {len(chains)} potential attack chains",
+            f"Highest probability chain: {max(chains, key=lambda x: x['overall_probability'])['overall_probability']:.2%}",
+            "Recommendations:",
+            "- Test chains in order of probability",
+            "- Prepare fallback methods for each stage",
+            "- Consider detection evasion at each stage"
+        ]
+
+        return "\n".join(recommendations)
